refactor: move generateUpgradeInfo to util/cli
